Move common test logic to FieldEncryptTestBase class.
